Changed puzzle rings again
Possible ring states: 1-5
When ring is at state 5 and button is pressed, it overflows back to state 1.

Each state is mapped to an angle of rotation: 1:72, 2:144, ..., 5:360

Rings follow this rotation, but don't even work somehow?? Feel free to spend either 5 minutes or 5 hours fixing this, because it is probably easier than I'm making it.